requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
coverage==5.4
deprecation==2.1.0
eyed3==0.9.6
feedgen==0.9.0
filetype==1.0.7
google-api-core==1.25.1
google-auth==1.25.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
libcst==0.3.16
lxml==4.6.2
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
